Fix constant condition elimination in sequence
We had some basic infrastructure for optimisation of constant
conditions, but this was almost completely hacked out, apparently in an
act of frustration. Fixing it turns out to be suspiciously simple. The
tests pass, at least.

Allowing the sequence handler to know when subexpressions are constantly
true allows it to generate much shorter code. The Parsoid tokenizer is
5% shorter after this change. There is not much change at runtime,
possibly because the JS compiler is doing a good job of propagating
constants.

Change-Id: Ibbc1a3b99841ce91943d8729c7213a08f8e8f6a4