Fix visitor for new multiple initializer feature
ast.initializer is now the same as before if there are 0 or 1
initializers, and an array if there are 2 or more. Update visitor for
this. During development I initially made it be an array
unconditionally, but I changed it to this more compatible format shortly
before merging in order to make the tests pass, but neglected to update
visitor.

Update version to 2.0.1

Change-Id: Icf665fd8aeff0f546555a803c8f5b8608e9a3c72