Simplify loops zero_or_more and one_or_more
Instead of generating the whole expression twice, use an infinite loop
that breaks from the middle. This generates compact code which is
probably efficient. The PHP bytecode dump confirms that it is better to
have one comparison than two. The expression will have a conditional branch
in any case, so piggybacking on that branch avoids the need for a second
conditional branch in the while().

Change-Id: I380f9793a8b1d8e6a7a7ef7491df5ce5f8c4151c