Fix caching of reference rule variables
If a cached rule set a reference variable, but happened to set it to
`undefined` (in JS) or `null` (in PHP), the value wouldn't be properly
restored when the cached value was reused.

While we're touching this code, use a name prefix instead of separate
`refs` array in order to avoid unnecessary allocation.

Bug: T221384
Change-Id: If4cf32f828ff8829bbaf3600de6cd853685a5376