Optimize silent fails (T224086)
Remove unnecessary failure-reporting checks that are always
false. A portion of this commit is based on Mingun's fix to issue #399
and #400 on PEG.js.

This avoids passing a 'silence' argument in the case where it can be
statically determined it will always be true, and avoids the repeated
`if (!silence) { peg$fail(...); }` where the condition will never be
taken.

Bug: T224086
Co-Authored-By: dan1wang <daniel.liberated@gmail.com>
Co-Authored-By: C. Scott Ananian <cscott@cscott.net>
Change-Id: Ib38af32cadc83c4c7be03b28d987e222879a2c9b