Added a place in the template to add header comments
The goal of this patch is to be able to add header comments pointing out
that the file that results in this is generated and should not be
edited. The default comment "Generated by WikiPEG" is taken from the
template.js. Scripts calling this library can change the comment by
passing it as a script parameter to spell out the exact way of
generating the output file, thus providing documentation to the
developers stumbling on generated files.

Change-Id: I074794b2e9ae56e434e6a8a809f5d62799f181be