Change namespace from WikiPEG to Wikimedia\WikiPEG
This matches the recommendation in PSR-4 that the first component of
the namespace is the "vendor namespace" and makes the namespace match
the composer package name.

Aliases are set up so that the old WikiPEG namespace will continue to
work.  In the next major release it is expected that the aliases
will be removed.

Bug: T287972
Change-Id: Ic235483e0fce6cc07a9e5c59328b99fe0129866f