Hook up runCommonTests.php to CI by running it in `composer test`
This only runs if `node` is available, which means it doesn't
currently run in the default image in Wikimedia CI.  It's still useful
for developers who run `composer test` locally.  In Wikimedia CI we
have a special "wikipeg-special-node10-plus-php72-composer-docker"
image with both node and PHP installed, and we run `make test`
(using the Makefile in this repository) in order to run both JS
and PHP tests.

Change-Id: I8f77d34edebdb6cfe91d787a9a25fce70fb86bf5