Bug fix for 'Rule reference variables should not be affected by failed rules'
When caching is used, the earlier patch could store an entry
in the cache which falsely recorded that a reference variable was
unchanged, due to the fact that the saved_* variable recording
the original value of the reference value was clobbered.

We need a stack of variables to store "original" reference
values, just like we have a stack of position variables,
to allow backtracking through nested sequences.  Uses the
new machinery to avoid creating too many new variables
when we can reuse the value stored by the parent node.

Bug: T222419
Follows-Up: I27599a7c3b177cb9443ac546d9539470e38c2af7
Change-Id: I60c8411c4da457611c3784f35ad6ed44c114b333