More aggressively throw away results during 'discard' mode
In particular, try hard not to hold onto strings and not to create
new substrings of the input if we don't need to.  Although matches
against literals are likely to be references to intern'ed strings,
computed substring results likely result in needless string
allocation, which can be quickly de-ref'ed in places where we
can't avoid the substring altogether.

This patch requires a major version bump, since a new helper method is
added to the PHP runtime's PEGParserBase.php.

This is a modest ~1% improvement in tokenization time on
[[en:Barack_Obama]] using Parsoid's wikipeg PHP tokenizer.

Change-Id: Ic3a05682ef1cc4157e5f0dfafcbf6a76189ae37f