Bug fix: register reuse in sequence evaluation
The logic for register freeing was mixed up in the sequence compiler,
with the result that it was possible for a later sequence result to
overwrite the result of an earlier one, causing duplicate entries
in the sequence array result.

The result.free list contains the result register and thus isn't
safe to free until the parent uses the result of the seq.  Because
we'd like to free the sequence element result values early, don't
merge them into result.free but instead save them in a separate
partsFree list and free them after we construct the result array
(but before we free the result register!).

Change-Id: I8a0e1bb1cb7609778cf59a1ca50574ac63398aed