Don't save position and ref state when it is reusable from parent
The first node in a sequence can reuse the saved position/ref state
used for restoring the sequence, and so doesn't have to allocate and
store its own.  Be careful about repetition nodes, however: their
child expression can't reuse the saved state from the parent because
even though this is correct for the first repetition, a failure on
the second repetition should reset only to the end of the last
successful match, not all the way to the start of the repetition.

Change-Id: I64c8f6e68dc8883766f9d4ae72c9ccd988a7c597