Analyze FIRST(n) for each node and optimize subrule calls
A new pass analyzes the FIRST set for each node: the set of characters
which can begin a valid match.  Perhaps more usefully: if the next
character in the input is *not* in FIRST, and the expression is not
nullable, the expression is guaranteed to FAIL.

This allows us to rewrite references to a non-nullable rule R as

  (&FIRST R)

which does a character class match before recursing into R.  This
allows us to fail fast if there is no way R is going to match, as
well as saving cache entries for R (and its subrules).

This does affect failure reporting: although the *position* reported
will be the same (since we are early-detecting failures on the first
character), the rule identified as the failing one could be parent
of the actual failing rule.  For this reason, the `noOptimizeFirstSet`
option is added to allow this optimization to be disabled where
precise failure reporting is required.  We also add a new
`--precise-failure` option to the `wikipeg` binary, and use it when
generating wikipeg's own parser.

Change-Id: I4abeb3ae93f62c52e070310b26b553658fe11279